+++ jest.config.js
@@ -12,3 +12,4 @@
    
     // ...other configurations
   };
+