+++ CHANGELOG.md
@@ -1,2 +1,8 @@
 # Changelog
 
+## [0.1.0] - 2023-12-05
+
+### Added
+- added package.json and updated dependancies in it.
+- added contributing, changelog, jest config, security, rollup config, ts config
+